Complement stylesheet for Light Style extension
By copying over relevant parts of the `stylesheet-light.css` from
Luminus Shell.

The reason for not using Luminus Shell directly is it wasn't as
performant as using the `Light Style` extension.

A variation could be to skip on using Light Style and just incorporate
the whole Luminus Shell stylesheet inside the `Dash to Panel`
stylesheet. That was also more performant than using the Luminus Shell
directly, but a little less stable than the chosen solution, it seemed to me.